adds basic error assertions
